Added new Enemy AI
Added "Dude with a Bat" and "Cop" Enemy AI. They will pursue the player indefinitely.

Updated Player prefab so "Orientation" is now tagged as "Player", this is so that the physics push from the melee enemy and bullets will work.
This rigidbody is cached in every enemy controller.
Made some animation changes.